requirements.txt
# ============================================================================
# UNIFIED LANGGRAPH CHATBOT - REQUIREMENTS
# ============================================================================

# Core LangChain & LangGraph
langchain==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.1
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.1
langgraph-sdk==0.2.0
langsmith==0.4.8

# Community Tools & Integrations
langchain-community==0.3.27

# OpenAI
openai==1.98.0

# Vector Store & Embeddings
faiss-cpu==1.9.0.post1  # Use faiss-gpu if you have CUDA
numpy==2.3.2

# PDF Processing
pypdf==5.1.0  # More reliable than PyPDF2
pdfplumber==0.11.4  # Alternative PDF loader

# Web Tools
requests==2.32.4
requests-toolbelt==1.0.0
duckduckgo-search==7.1.2  # For DuckDuckGoSearchRun
wikipedia==1.4.0
beautifulsoup4==4.12.3

# Streamlit UI
streamlit==1.47.1
streamlit-chat==0.1.1  # Optional: Enhanced chat components

# Database
aiosqlite==0.21.0
SQLAlchemy==2.0.42
sqlite-vec==0.1.6

# Environment & Configuration
python-dotenv==1.1.1

# Data Processing
pandas==2.3.1
pyarrow==21.0.0

# Async Support
asyncio==3.4.3
aiofiles==24.1.0

# MCP Support (Optional - install if using MCP)
# langchain-mcp-adapters==0.1.0

# Monitoring & Debugging
tenacity==9.1.2
tqdm==4.67.1

# Serialization
pydantic==2.11.7
pydantic_core==2.33.2
orjson==3.11.1

# Type Checking
typing_extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
click==8.2.1
colorama==0.4.6

# Data Visualization (for Streamlit)
altair==5.5.0
pillow==11.3.0

# HTTP & Network
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
certifi==2025.7.14
idna==3.10
charset-normalizer==3.4.2
urllib3==2.5.0
sniffio==1.3.1
anyio==4.9.0

# Templating & Markup
Jinja2==3.1.6
MarkupSafe==3.0.2

# JSON Processing
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
referencing==0.36.2
rpds-py==0.26.0

# Token Processing
tiktoken==0.9.0
regex==2025.7.31

# Streamlit Dependencies
blinker==1.9.0
cachetools==6.1.0
tornado==6.5.1
watchdog==6.0.0
protobuf==6.31.1
narwhals==2.0.1
pydeck==0.9.1

# Git (for version control)
GitPython==3.1.45
gitdb==4.0.12
smmap==5.0.2

# Other
attrs==25.3.0
distro==1.9.0
greenlet==3.2.3
six==1.17.0
toml==0.10.2
xxhash==3.5.0
zstandard==0.23.0
packaging==25.0
ormsgpack==1.10.0
jiter==0.10.0
PyYAML==6.0.2
tzdata==2025.2

# ============================================================================
# OPTIONAL ENHANCEMENTS (Uncomment to use)
# ============================================================================

# Enhanced Search
# tavily-python==0.5.0  # Better than DuckDuckGo for research

# Better PDF Handling
# pdfminer.six==20231228
# pdf2image==1.17.0

# Image Processing
# Pillow==11.3.0

# Audio/Video Processing
# whisper==1.1.10  # OpenAI Whisper for transcription

# Redis for distributed caching (production)
# redis==5.2.2
# hiredis==3.0.0

# PostgreSQL support (alternative to SQLite)
# psycopg2-binary==2.9.9
# langgraph-checkpoint-postgres==2.0.0

# Monitoring & Observability
# prometheus-client==0.21.0
# sentry-sdk==2.20.3

# Rate Limiting
# slowapi==0.1.9

# Testing
# pytest==8.3.4
# pytest-asyncio==0.25.2
# pytest-cov==6.0.0

# Development Tools
# black==24.10.0
# flake8==7.1.1
# mypy==1.14.1
# pre-commit==4.0.1

# ============================================================================
# NOTES
# ============================================================================
# 1. For GPU support with FAISS, replace faiss-cpu with faiss-gpu
# 2. For production, consider using PostgreSQL instead of SQLite
# 3. Install MCP adapters only if you plan to use MCP servers
# 4. Use virtual environment: python -m venv venv && source venv/bin/activate
# 5. Install with: pip install -r requirements.txt
